// frontend/lib/generated-api/models/WorkflowRun.ts
/* tslint:disable */
/* eslint-disable */
/**
 * AI Analyst API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { WorkflowRunStatus } from './WorkflowRunStatus';
import {
  WorkflowRunStatusFromJSON,
  WorkflowRunStatusFromJSONTyped,
  WorkflowRunStatusToJSON,
  WorkflowRunStatusToJSONTyped,
} from './WorkflowRunStatus';
import type { WorkflowRunType } from './WorkflowRunType';
import {
  WorkflowRunTypeFromJSON,
  WorkflowRunTypeFromJSONTyped,
  WorkflowRunTypeToJSON,
  WorkflowRunTypeToJSONTyped,
} from './WorkflowRunType';

/**
 *
 * @export
 * @interface WorkflowRun
 */
export interface WorkflowRun {
  /**
   * The unique identifier for the workflow run
   * @type {string}
   * @memberof WorkflowRun
   */
  id: string;
  /**
   * FK for the project that this workflow run belongs to
   * @type {string}
   * @memberof WorkflowRun
   */
  projectId: string;
  /**
   * The type of the workflow run
   * @type {WorkflowRunType}
   * @memberof WorkflowRun
   */
  type: WorkflowRunType;
  /**
   *
   * @type {string}
   * @memberof WorkflowRun
   */
  langgraphThreadId: string;
  /**
   * The status of the workflow run
   * @type {WorkflowRunStatus}
   * @memberof WorkflowRun
   */
  status: WorkflowRunStatus;
  /**
   * The timestamp when the workflow run was created
   * @type {Date}
   * @memberof WorkflowRun
   */
  createdAt: Date;
  /**
   * The timestamp when the workflow run was last updated
   * @type {Date}
   * @memberof WorkflowRun
   */
  lastUpdatedAt: Date;
}

/**
 * Check if a given object implements the WorkflowRun interface.
 */
export function instanceOfWorkflowRun(value: object): value is WorkflowRun {
  if (!('id' in value) || value['id'] === undefined) return false;
  if (!('projectId' in value) || value['projectId'] === undefined) return false;
  if (!('type' in value) || value['type'] === undefined) return false;
  if (!('langgraphThreadId' in value) || value['langgraphThreadId'] === undefined) return false;
  if (!('status' in value) || value['status'] === undefined) return false;
  if (!('createdAt' in value) || value['createdAt'] === undefined) return false;
  if (!('lastUpdatedAt' in value) || value['lastUpdatedAt'] === undefined) return false;
  return true;
}

export function WorkflowRunFromJSON(json: any): WorkflowRun {
  return WorkflowRunFromJSONTyped(json, false);
}

export function WorkflowRunFromJSONTyped(json: any, ignoreDiscriminator: boolean): WorkflowRun {
  if (json == null) {
    return json;
  }
  return {
    id: json['id'],
    projectId: json['project_id'],
    type: WorkflowRunTypeFromJSON(json['type']),
    langgraphThreadId: json['langgraph_thread_id'],
    status: WorkflowRunStatusFromJSON(json['status']),
    createdAt: new Date(json['created_at']),
    lastUpdatedAt: new Date(json['last_updated_at']),
  };
}

export function WorkflowRunToJSON(json: any): WorkflowRun {
  return WorkflowRunToJSONTyped(json, false);
}

export function WorkflowRunToJSONTyped(value?: WorkflowRun | null, ignoreDiscriminator: boolean = false): any {
  if (value == null) {
    return value;
  }

  return {
    id: value['id'],
    project_id: value['projectId'],
    type: WorkflowRunTypeToJSON(value['type']),
    langgraph_thread_id: value['langgraphThreadId'],
    status: WorkflowRunStatusToJSON(value['status']),
    created_at: value['createdAt'].toISOString(),
    last_updated_at: value['lastUpdatedAt'].toISOString(),
  };
}